requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
insightface==0.7.3
onnxruntime==1.17.1
cvzone==1.6.1
flask==3.0.3
pillow==10.2.0
appwrite==15.3.0
python-dotenv==1.0.1
